+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 [1.2.6] - 2018-11-08
 * new option: orientation
 * new option: disableWeekends
+* fixed bug #23
 
 [1.2.5] - 2018-09-06
 * Improve pluralize. Now as function.